requirements.txt
Flask
tensorflow
requests
numpy
gunicorn
imageio
flask-bootstrap
dash
Pillow